requirements.txt
cryptography
wechatpy
PySocks
Flask
furl